Update `plugin.mustache` and `plugin-packages.mustache`
Currently if you scaffold a child theme and a plugin, the plugin headers
are formatted differently than the child theme headers. The plugin also
uses a different version number `0.1-alpha`), which is itself different
than the version used in the plugin's `package.json` (`0.0.0`). This
commit normalizes the header formatting and updates the version numbers.